Ensure integer when calling chunks

+++ km3pipe/utils/qrunqaqc.py
@@ -178,7 +178,7 @@ class QAQCAnalyser(object):
         else:
 
             if batch_size is None:
-                batch_size = math.ceil(total_runs_to_process / max_jobs)
+                batch_size = int(math.ceil(total_runs_to_process / max_jobs))
 
             print(
                 "Proceeding with {} runs distributed over {} jobs, {} runs/job"
